requirements.txt
torch
transformers
datasets
sentence-transformers
faiss-cpu
streamlit
pandas
numpy
fastapi
uvicorn
tqdm
